Change fetching events to an observable operation
Instead of directly launching the events api request, the request should
be triggered from the `FETCH_EVENTS` dispatched action.

Navigating to the events page dispatches a navigation action, and one of
the epics translates the first navigation into a `FETCH_EVENTS` action.

Because the api request is now handled by observables, there is no
direct way to interact with the api call. To show a status message
during the fetch, the `isFetching` variable is moved to the redux store.