Implement a request module for api request
To handle logic around making api requests, the request module will
handle the execution of making the requests instead of having the logic
spread around the application. Overall, this is a work-in-progress of
logic separation and encapsulation and is expected to mature.

The `Conn` struct serves as the single encapsulation point for all
parameters around an api call.

The `Service` struct describes the primary connection to an api service.

The `Resource` struct describes a single resource of a service and
includes the allowed actions. The export of a resource should include
constructed api request calls to simplify consumers.

The pulic `request` takes a Conn or builder and constructes an api call
to the described service and resource.