Create a standardized snackbar for errors
To standardize the notification of errors, the errorSnackbar epic will
watch for error actions that have a message property in the meta object
and dispatch an action to show the messsage on the ErrorSnackbar. The
action to show a message is also watched for to later send an action to
clear the message.

While the snackbar is active, other errors are ignored because,
presumadly, they are a result of the current error being shown.

Potentially, an error action with a message may not want to display in
the ErrorSnackbar. Other flags could be added to the meta object for
filtering.