Disable and unregister service worker
The service worker is causing the undesired caching. Unregistering it
will allow for expected non-caching until a desired way to make use of
it is found.

As a visible stop-gap, a confirmation will prompt for reloading the page
to load new content if the user still have the service worker
registered.