Implement apiRequestBuilder
Instead of building types, actions, and epics for api requests by hand,
the apiRequestBuilder can orchestrate all of the details.

As an intermediate step, the same exports are being kept, and only the
data module internals are changed.

The errorCreator and errorSnackbar is being adapted to account for 400
vs 500 error messages.