Burn everything down and rebuild
By planning for future features in the architecture design, the codebase
became too complex. The work is still in git history, but everything is
being stripped down to basic react to reflect the current state of the
project. Parts of the design could return as the website grows.

The simplification is partly driven by the desire to make onboarding
easier. The friction should be lessened while there is no directe need
for it.